FilterQuery : Fix context used to hash the input filter
We need to provide the scene so that the filter can use it in computing its hash. I'm not sure why `SetFilter::hashMatch()` is a no-op if the scene isn't provided - it might be better if the Filter base class threw an exception if the scene wasn't specified so problems like this can't go undetected. I'm not sure what the consequences of that might be though - might be best to tackle that as part of a larger Filter API overhaul.

Fixes #5066.