ImageView : Fix display transform regressions
The node for the "Default" transform registered by `startup/gui/ocio.py` is modified on the fly as users edit the preferences, so we need to respond to `plugDirtiedSignal()` and regenerate the shader.

The same preferences mechanism also allows users to specify OCIO context variables, which may themselves reference Gaffer context variables from the global context. So we need to scope that context when getting the processor, and make sure we dirty the shader when the context changes. The vast majority of context changes will be irrelevant though, and generating a shader isn't free. So we store a hash of the processor to avoid redundant repeat shader generation.