ArnoldOptionsUI : Add user defaults for GI depth plugs
The intention with the ArnoldOptions plugs is that they should have the same default values as Arnold itself. So you only need to enable one and change it if you don't want the Arnold default. But Arnold's default for `GI_diffuse_depth` and `GI_specular_depth` is `0` and not `2`, so we messed up at some point.

We can't change the actual default value of the plugs, because that will break old serialisations (because we omit `setValue()` calls when plugs are at the default value). But we can register `userDefault` metadata so that newly created nodes will be given the proper "default" via a `setValue()` call.