LabelPlugValueWidget/PlugLayout : Fix "green dot" for compound plugs
It wasn't correctly considering `userDefault` metadata or input connections on child plugs.

The "todo" in the previous code implied that we should have pushed all this new logic into `NodeAlgo.isSetToUserDefault()`, but I didn't do that. That function is intended to mirror `ValuePlug.isSetToDefault()`, and therefore allows connections from promoted plugs and the like (where the incoming value is non-context-sensitive). As documented in the new LabelPlugValueWidgetTest, this is not what we want for the UI, where we are considering whether or not any edit has been made by the user. So the new logic is instead implemented on LabelPlugValueWidget directly.