MenuButton : Fix missing arrow indicator in PresetsPlugValueWidget
This was broken in 5b22c8fcd9e96bbc88d833c81a60c51e4fac7426, because it moved the first `setText()` call _after_ the point the widget was first shown and polished. After that, if we want any Qt property changes to be picked up by the stylesheet, we must do a manual repolish.