ShadingEngine::{anonymous}:convertValue : Support testing if allowed
The comment on convertValue says "Equivalent to `OSL::ShadingSystem:convert_value()`".
This means it is reasonable to expect that it should also support this feature of
convert_value:
"""
As a special case, if dst==NULL or src==NULL, no copying is performed, and
convert_value merely returns a bool indicating if the proposed type conversion
is allowed.
"""