ImageView : Fix display transform update bug
This manifested itself as it being impossible to switch back to a previously used display transform in the Viewer. It was introduced in 549726641a6e4448d3d8f3a1d6e86263ba78c0a2, where I failed to realise that we were previously setting the post process shader unconditionally before _every_ frame, and that was necessary in order to process any new assignments to `m_displayTransformAndShader`. This seems wasteful, but this commit goes back to that approach as its the easiest way of fixing the bug without rejigging things on a larger scale.