LightEditorTest : Fix testing w/ multiple columns
- Getting any old `_GafferSceneUI._LightEditorInspectorColumn` will also
include the new `_GafferSceneUI._LightEditorSoloColumn`, which we don't
want.
- It turns out creating a Python list with multiple entries using the
`[] * n` syntax creates shallow copies of the list contents. Therefore
when modifying the selection, it was affecting all columns because the
list members were all the same `IECore.PathMatcher()`. This caused
incorrect results when both the "Mute" and "Solo" column were toggled.