SceneGadget : Fix crash switching from Cycles to OpenGL
The crash was caused by the CyclesRenderer sending pixels to the OutputBuffer _after_ `CyclesRenderer::pause()` was called. This meant they got sent to an OutputBuffer that we had already destroyed. None of our other renderer backends do this, and I think there's a definite case that `CyclesRenderer::pause()` is at fault here.

I did try an alternative implementation of `CyclesRenderer::pause()` that blocks until all pending pixels have been sent, and it fixed the crash. But it  needs careful consideration and testing (some updates were faster with it, but some slower, and it has implications for the rest of the CyclesRenderer implementation).

So instead, I made the fix you see here. `CyclesRenderer` does at least make sure all remaining pixels are sent during its destruction, so as long as we keep the OutputBuffer alive until that is done, we're OK.

Fixes #5051