CI : Test GafferArnold and GafferArnoldUI
The addition of `powershell -c` to the Windows `testRunner` parameter
is needed because `Invoke-Expression` is a Powershell command.
Powershell is the default shell for github actions, so we didn't need
it to run the tests directly from the GH Actions command. But running
via Python `subprocess`, the shell is given by the `COMSPEC` environment
variable, which is `cmd.exe`. So we call Powershell regardless which
allows us to use `Invoke-Expression`.