Backups : Fix bug loading non-ASCII backups if locale was changed
If we don't specify the encoding to `open()`, then Python chooses a default encoding which is platform-specific. On Linux, this is generally UTF-8 but is unfortunately sensistive to a bug in OpenShadingLanguage which changes the locale when loading a shader (see #5048). On Windows, UTF-8 is not even the default encoding, so the new test case may have failed even without the explicit choice of locale. This is described pretty well in [PEP 597](https://peps.python.org/pep-0597).

The solution is simply to specify the encoding when opening the file, via the `encoding` argument to `open()`. But that argument only exists in Python 3, so we must use `io.open()` instead, to bridge the gap until we drop Python 3 support. Eventually, we can even drop the `encoding` argument entirely, because [PEP 686](https://peps.python.org/pep-0686/) makes UTF-8 the default on all platforms, and is adopted in Python 3.15.