Specify encoding for all calls to `open()`
As documented in [PEP 597](https://peps.python.org/pep-0597/), calling `open()` without specifying an encoding is not cross-platform-friendly. And as shown in #5048, it even causes problems on Linux in an environment where we think the locale is UTF-8 friendly. So we now always explicitly pass `encoding = "utf-8"` to every call to `open()`. This matches the [default behaviour in Python 3.15](https://peps.python.org/pep-0686/), so one day we'll be able to rip it all back out again.

This is a Python-3-only companion to PR #5126, which fixes this specific issue for `Backups` in a way compatible with Python 2. I deemed that approach to be too fiddly to apply across the board when we are so so close to leaving Python 2 behind forever.

Fixes #5048.