PresetsPlugValueWidget : Cache current preset
We were calling `NodeAlgo.currentPreset()` from the menu definition, which could trigger a blocking compute that could hang the UI indefinitely. Instead we store the current preset in `_updateFromValues()` and use the stored value
when building the menu. This also means that the menu state will be in sync with the state of the displayed text.

Also refactored the `isCustom` logic into a function to avoid repetition, and documented the possible values for `__currentPreset` (these are unchanged, but were undocumented before).