PresetsPlugValueWidget : Avoid layout flicker
This was most obvious in the Viewer toolbar when viewing a scene for the first time - the layout jumped up and down before settling at the top. The issue was that when first shown, the custom widget was visible and therefore claiming space. Then when we hid it, it didn't need the space and the layout collapsed to the needed size.

Because `self.__currentPreset` isn't up to date in the constructor, the visibility will be incorrect in one circumstance - when all widgets allow custom values, _and_ the `isCustom` metadata isn't set on any of then, _and_ none of the values match a preset. This is exceedingly unlikely, so I think a tiny bit of flicker would be acceptable in that case.