SceneGadget : Destroy camera before renderer
It's an unfortunate requirent of the Renderer API that all ObjectInterfaces must be destroyed before the Renderer they came from. By not destroying the camera before the renderer, we ended up calling `AiNodeDestroy()` _after_ `AiUniverseDestroy()`, which has already implicitly destroyed all the nodes. That caused crashes when shutting down the Arnold renderer in the Viewer.

This bug was introduced in #5120, ironically when fixing a Cycles shutdown crash. Since that hasn't been released yet, we don't need a specific mention of this latest fix in Changes.md.