GLWidget : Remove context sharing with Maya and Houdini
This isn't the last word on the subject, because we do need to continue to support this somehow for Image Engine. But we also want to start using QOpenGLWidget instead of QGLWidget (which is deprecated already and is removed in Qt 6), and QOpenGLWidget doesn't support the sharing mechanism we're using. So the plan : make a minimal and clean implementation, and then figure out what we need to do to re-enable hosting in Maya and Houdini.