GLWidget : Set minimum size of one pixel
This stops Qt trying to make framebuffers with zero size, which causes a warning to be printed, and makes the draw fail, requiring a second kick to get the GLWidget up to date. See https://github.com/GafferHQ/gaffer/pull/5173#issuecomment-1447494169 for fuller description of the symptoms.

In practice, the minimum size doesn't have any impact on screen real-estate, because our GLWidgets are usually in a SplitContainer within a layout, and that ignores the minimum size when allocating space. So we can still maximise the other widgets and/or drag the splitter, thus obscuring the GLWidget completely. It's just that the obscured widget still has enough size to prevent the Qt error.