Arnold Viewer : Delete `ai:fis_filter` option
This is an experimental option that we don't expose officially in Gaffer, but which can be enabled via a CustomOptions node. It overrides all the filters for all AOVs in the scene. This replaces the `closest` filter we use for the Viewer's ID render with a `box` filter, which promptly complains that it doesn't support integer data and then refuses to send any data. The upshot : you can't select anything in the Viewer, and if you select anything elsewhere then we render _everything_ with the blue selection highlight.