StandardStyle : Fix OpenGL blending functions
Previously we were applying `GL_SRC_ALPHA` to the alpha channel as well, resulting in `alpha * alpha` being written to the framebuffer. This would break any subsequent operations relying on that alpha.

Rather than try to save and restore yet more state in `ViewportGadget::renderInternal()`, we simplify things by rebinding the style at the start of each layer, allowing it to clobber everything.