SConstruct : Verify preferred encoding
This defaults to something very unhelpful on Windows, hence the previous workaround that monkey-patched in the UTF8 encoding. That workaround was sufficient in Python 3.7, but does not work in Python 3.10, which is what my SCons install happens to use. I've been unable to find a similar hack for 3.10, and since setting PYTHONUTF8 is so straightforward, it seems reasonable to ask everyone building on Windows to do that, regardless of Python version. This also gives us a little bit of a safety net on Linux, where the preferred encoding has always been UTF8 so far, but plausibly could differ on someone's machine somewhere.