Docs : Work around excessive delays capturing screenshots of renders
It appears that we're getting into something of an "event storm", where we don't get to an idle event because new pixels arrive during a draw, causing another draw event. My assumption is that the new overhead of sampling/blitting in the ViewportGadget has pushed Mesa over a threshold where there's time for the renderer to get new pixels in before we've completed event processing. We saw a similar problem when we had CI on MacOS (using the original ViewportGadget implementation) and had to disable the documentation generation there. By not drawing the UI while waiting for the render, and stopping the render before taking the screenshot, we can avoid this situation entirely.