StandardLightVisualiser : Don't make unnecessary `Groups`
Not all groups ended up having anything in them. And groups with a single child were unnecessary for many visualisations because they didn't require a transform for orientation.

The one downside of this change is that in some cases we're creating two visualisations instead of one. But in all cases those are going to be need to be split apart soon anyway, because each visualisation needs a different colour transform.