ImageGadget/ViewportGadget : Rejig floating point texture reporting
The old message no longer made much sense, because the ImageGadget isn't responsible for applying the display transform any more, and the CPU mode is no longer a thing. Reworded and moved the message to ViewportGadget, where we have greater reliance on floating point textures, since this is where we're now applying the display transform.

Of course, ViewportGadget is used in far more tests than ImageGadget, so the message will now be emitted more. So instead of polluting loads of tests with CapturingMessageHandler, I added a `TestCast.ignoreMessage()` method and used it to ignore the warning once in the `GafferUITest.TestCase` base class.