UI TestCase : Test NodeGadget lifetimes
In practice NodeGadgets are typically implemented in C++, so it's unlikely they would have the lifetime issues we're looking for here. But it's still useful to construct one for every node to ensure there are no other problems - this uncovers an issue whereby NoduleLayout registrations for the OSLObject and OSLImage nodes have gone missing on Windows.