Avoid deprecated `boost/bind.pp`
We tidied this all up previously in 72b0e8eff9f1bf1dbf3011affdbd20dc739515fe, using `boost/bind/bind.hpp` instead of the deprecated `boost/bind.hpp`. But problems crept back in because of the workaround (for Boost itself!) we had in the SConstruct. We can now remove that workaround for Boost 1.80, and keep our code clean.