View/OpenColorIOAlgo : Remove `unpremultiply` parameter
It was not being used, and always had a value of `false`.

It took me a while to figure out why the `false` value was correct, since actually the contents of the texture we're rendering _are_ premultiplied, so you might think we need to unpremultiply before applying the transform. This is _not_ the case though, because conceptually we're transforming the _rendered image_, not the surface texture before coverage is taken into account (using 3d renders as an example).