SpreadsheetUI : Use Qt's alternating row colours
Now that rows can be hidden from the UI based on a row name filter pattern, the previous method of alternating row colour based on the row index could result in the same colour being applied to two adjacent visible rows. Qt's approach accounts for row visibility.

The addition of the `gafferReverseRowColors` property allows the use of `$backgroundAlt` for the regular item and `$background` for the alternate. This is required to maintain visual consistency with the previous approach, keeping the row colour of the default row distinct from the first rows of the rowNamesTable and cellsTable.