Documentation : Adjust image annotation offsets
Adjusted the annotation in interfaceSpreadsheetNodeBreakdown to better fit the new default label position in 7423abe4.
Adjusted the placement of the card masking the add row button in interfaceSpreadsheetNodePatternWidths, as the button placement has changed in e8dfa56a. Also remove a now unnecessary card that was previously masking a scrollbar.