Headers : Prefer `fmt::format` to `boost::format`
This means that practically all the libraries need to link against `libfmt`, so instead of listing it explicitly for each one, I've just moved it to the standard template environment in the SConstruct.

The one annoying thing here is the change in ContextTest : `format()` refuses to format pointers unless they are void pointers, so we need to cast.