ImageReaderTest : Adapt `testColorSpaceOverride` for ACES configs
There appears to be no direct equivalent of the old `Cineon` color space in ACES. The closest in intent seems to be `ADX10`, which _is_ a log space, but seems to have a completely different range to `Cineon`. Rather than generate more binary blobs in the repo, it's easier just to test directly against a color transform.