Image(Reader|Writer)Test : Account for switch to ACES OCIO config
These tests compare files that were written with Gaffer's previous default OCIO config, where all colorspaces shared the same primaries (assumed to be sRGB) and the only difference was the encoding. Now we are using ACES, the 8-bit files continue to be treated as sRGB but the EXR file would by default be assumed to have ACEScg primaries. Rather than rewrite all the example files, its currently simpler to tell the ImageReader that the EXR file uses sRGB primaries too (which was the intention at the time it was created).