Spreadsheet docs : Use `scene_linear` role instead of `linear` colorspace
The former is available in the ACES OCIO config, whereas the latter is not, which was causing the documentation build to fail.

It seems that all the ImageReader and ImageWriter nodes here are dealing with PNGs, but were claiming that they were in linear space, effectively making Gaffer's working space be `sRGB`. I assume this was so that the overlay colours could be specified in this space, but I don't know. I suspect there's a better way of doing all this, but that's for another day.