OCIO config : Remove "Default" DisplayTransform
As requested in #5215. This does leave the preferences plugs rather badly named - they're now just controlling the colorspace used to display swatches etc in the UI widgets, and not the main display transform. Should we rename them? Or are we going to end up removing them in favour of per-script settings anyway?