Cycles : Tag tangent primitive variables
This allows normal maps to be rendered using the procedure described in #5269. It's really just a stop-gap measure until we (or ideally Cycles) can generate tangents on demand based on the needs of the assigned shaders. It also differs from #5269 in that we require the names to be exactly `uv.tangent` and `uv.tangent_sign` rather than just ending with `tangent_sign` or `tangent`.