Fix MSVC warning C4702
C4702 warns against unreachable code.

The change in `LRUCacheTest.cpp` is an attempt to work around an odd
warning, maybe a bug in MSVC. Originally, the warning about unreachable
code was coming from the `return 0;` statement. But removing that
created a new error stating that `void can't be converted to an
integer`. By adding a conditional, `return 0` can potentially be
reached, although the tests will always throw because they supply a
non-zero value to `key`.