Disable MSVC warning C4723
C4723 warns against a possible divide by zero. Rather than disable the
warning across the whole project, we disable it in the one location
it is currently being raised.

The longer-term solution is to prevent the divide by zero from happening
in `OpDivide`, but we don't have a clear idea of what the proper result
of a divide by zero should be at the moment.