AngularDrag : Fix `PlanarDrag` axis setup.
Previously, the comment describing the `AngularDrag` constructor stated that the axis at which rotation would be zero was `axis1`. The implementation made `axis1` the direction of 90 degrees rotation instead. Changing the argument names to `normal` and `axis0` makes their meanings clearer and we can also remove the negation of the cross product between (formerly) `axis0` and `axis1` and a cross product calculation to get the drag direction.