Revert "PlugValueWidget : Don't unlock plugs that are statically locked"
This reverts commit c96c497131957b05aa71cd3c114807fa1f622b75.

I originally made the change to prevent users from unlocking the `Shader.name` plug which is locked statically in `ShaderUI.py`. But it caused consternation at Cinesite, where other plugs are locked statically as a suggestion to users that they shouldn't use them, but then those users insist on needing to unlock the plugs and use them anyway.

The downside of this is that it is now possible to unlock `Shader.name` and change it, at which point it will be out of sync with the shader loaded via `Shader.loadShader()`. But our answer for now is "don't do that".

The other thing that benefited from the reverted behaviour is the CameraTweaksUI, where `readOnly` was registered statically to prevent users changing the name of the parameter being tweaked (for the preset parameters). Longer term we should do what we do for NameValuePlugValueWidget, and use metadata to request a plain LabelPlugValueWidget rather than showing the `name` plug.