EditScopeUI : Hide `BoxIn.name` and `BoxOut.name` plugs
We don't want the user to edit their values, and hiding them completely adds a belt to the braces provided by `readOnly` metadata. We're not losing anything of great value by not displaying them in the NodeEditor, since the same information is available on the NodeGadget in the GraphEditor, and is always the same for these particular nodes.