Fix compilation with Clang
This was broken by 4e2786af7dfa0e2a823d890ab6d8e522ddc696ab, where we introduced a `GraphComponent::nameChanged()` virtual method which conflicted with several pre-existing methods with the same name.

In the case of Catalogue and Shader we can override the new virtual function instead of connecting to signals. In the case of NameGadget and DotNodeGadget we just rename the functions so they don't conflict.