SceneReader : Check `setNames` before loading a set
Otherwise `IECoreUSD::USDScene` will do a full stage traversal looking for set members, only to return with nothing. This can represent a huge overhead when many sets exist in the Gaffer scene, but not in the USD file. This reduces the runtime for `testUSDEmptySetsPerformance` by greater than 85%. I suspect it will have an even greater impact on the production scene I'm currently testing.

I did consider moving this check to `USDScene::readSet()` itself, but decided against for a few reasons :

- `readSet()` can be called at any level of the hierarchy, so we would either need a special case for the root (the only case Gaffer cares about), or we would
need storage for the set names at each location (to cater to the general case).
- If we need to store things somewhere, using Gaffer's single unified cache is preferable to an invisible cache somewhere else.
- The same issue applies for `AlembicScene`, so this way we get to kill two birds with one stone.