SceneReader : Use GlobalScope for accessing input plugs
This prevents impossible setups like requesting a different filename for each location. It also reduces pressure on the hash cache in the common case that the filename is provided by an expression input (which would previously have been hashed once for every value of `scene:path`).

We originally tried this in #1745, but abandoned it due to concerns that the additional context creation overhead offset any benefits. But Context creation performance has been significantly improved since then. I tested this on the Moana island asset, _without_ an expression driving the `fileName` - the worst case situation where there can be no benefit from the GlobalScope. This demonstrated that the additional overhead of constructing the GlobalScope is now neglibible.