VisibleSet : Rename `match()` to `visibility()`
It was proving cumbersome to fit all of the specifics of a VisibleSet test into an `IECore::PathMatcher::Result` so we now replace it with a `VisibleSet::Visibility` struct and rename `match()` to `visibility()` as that makes more sense given the new return type.

This allows us to separate the path's `drawMode`:

- `None` : Should not be drawn.
- `Visible` : Should be drawn.
- `ExcludedBound` : Should be drawn but only as a bounding box signifying that the path has been excluded.

From whether the path could have `descendantsVisible`, which is now used to instruct the RenderController as to whether the path should be expanded.