ColorProcessor/OpenColorIOTransform : Refactor for simplicity and performance
Derived classes now return a ColorProcessorFunction which is cached on an internal `__colorProcessor` plug. This simplifies the implementation of pass-throughs for no-op processors (they just return an empty ColorProcessorFunction and the rest is dealt with in the base class). It also means that we only create OCIO CPUProcessors once, instead of every tile. This gives a 25% reduction in runtime for a test applying a ColorSpace node. Tests with `gaffer stats -image -preCache` also show a significant reduction in hashing time - around 65%.