RenderController : Use `VisibleSet::Visibility` to control visibility
`VisibleSet::Visibility::drawMode` informs us of the draw mode of a particular path.

Paths with a `drawMode` of `None` are not drawn, but are still expanded when `VisibleSet::Visibility::descendantsVisible` is `true` to allow one or more of their descendants to be drawn.
Paths with a `drawMode` of `Visible` are drawn.
Paths with a `drawMode` of `ExcludedBounds` are only drawn as a Placeholder bounding box.