Color widgets : Use `Widget.displayTransform()`
Also remove `useDisplayTransform` option, because it no longer makes sense - clients can just call `setDisplayTransform( identityDisplayTransform )` instead.

There's a change in behaviour here for ColorSwatch, because it used to apply the display transform to the background colour when `useDisplayTransform` was off. We can't really do that any more, because when the display transform is identity, we don't know what it would have been otherwise. Possibly we could see if the parent has a different display transform and use that, but that would be a gigantic guess. In practice, none of this seems to matter, because the places we used it are showing RGB colours without an alpha.