Arnold Renderer : Fix `step_size` handling
We were completely ignoring `step_scale` if `step_size` was explicitly set to `0.0` (meaning auto). This was different to the results when `step_size` was left unspecified, despite it ostensibly defaulting to `0.0`! This is nasty, because it thwarts attepts at render optimisation, and it takes some fairly detailed reading of the Arnold stats to detect the problem.

Because this fix can change the appearance of Arnold renders, we should carefully consider where it belongs. The possible hardline stances are :

- This is a bugfix and should go into `1.1.9.x`.
- This is a behaviour change and should go into `1.3.0.0`.

I've taken a middle ground by putting it into `1.2.x.x` and adding a caution to the release notes. I definitely don't think it's suitable for `1.1.9.x` because shows in that phase of "warts and all" version lock won't want _any_ changes to images. I could be persuaded pretty easily to move to `1.3`, but I know of shows starting up in `1.2` shows that could benefit from the increased clarity when optimising their renders.

The original bug dates back to PR #2634, which contains some useful information on our original intentions and the reasons for them, which I've distilled into the new comment in the source code. I've also verified that the Arnold behaviour hasn't changed since what we observed back then.