OpenColorIOConfigPlugUI : Register display transform before use
In typical use we expect `OpenColorIOConfigPlugUI.connect()` to have been
called for any ScriptNode which has a `View` created for it, but that is not
the case when running the unit tests, because those create ScriptNodes that
aren't parented to any `Application`. One such example is `GafferSceneUITest.SceneViewTest`,
which actually displays the UI when the tests run, and therefore ends up invoking `DisplayTransformPlugValueWidget.__ensureValidValue()`.
This was then changing to a display transform that hadn't been registered yet
(because `connect()` wasn't called). The solution is to register a display
transform just in time, before we set the plug value.

This only appeared as an issue because `GafferUITest.LayoutsTest` runs the
`gui` startup files, causing any later tests to run with OCIO configured.
And it only showed up on Windows CI because that runs the tests in a different
order to Linux. It was therefore reproducible on either platform like so :

`gaffer test GafferUITest.LayoutsTest GafferSceneUITest.SceneViewTest.testLookThrough`

I think there's a good argument that LayoutsTest
shouldn't be polluting the tests with the `gui` configs, but I've robustified
OpenColorIOConfigPlug anyway, because I think future "context pinning" features
could potentially place us in a similar situation.