Context : Fix crash triggered by re-entrant call to `set()`
We were updating `m_allocMap` (which maintains ownership) _after_ assigning to `m_map` (which deals in raw pointers) and emitting `changedSignal()`. This meant we were in an inconsistent internal state at the point we triggered arbitrary observer code via slots connected to the signal. If a slot set the _same_ variable again, we'd end up with the raw pointer from the second call to `set()`, but with the owning pointer for the first call. Hence we were referencing a dangling pointer and subsequent attempts at accessing it would crash.

The solution is to store to `m_allocMap` _before_ calling `internalSet()`, so everything is in sync when we emit the signal.

This fixes crashes triggered by the interaction between multiple HierarchyViews synchronising via `ContextAlgo::set/getVisibleSet()`, as described in (#5364). I'm still not 100% sure how that was leading to a re-entrant call, but it seems worth fixing this at the lowest level possible, and "no signalling without consistent internal state" is a rule to live by anyway.