NumericPlugValueWidget : Fix increment / decrement
This fixes a bug in a specific case of using the increment / decrement
keyboard shortcuts. If the plug value was changed with the keyboard
shortcuts in a way such that the number of digits in the value changed
and the cursor was at or adjacent to a digit place that was removed,
it would be reset to the end.

This was the result of :
1. The widget incremented the value
2. The plug was updated from the widget
3. The widget was then updated from the plug

The final update in step three updated the plug from the numeric value,
which does not have the zero-padding intended to keep the number of
digits consistent during the increment sequence. Here, we prevent that
value from being updated a second time in the case of an increment
operation.