NumericPlugValueWidget : Fix animated value change
Previously, incrementing an animated numeric plug with the keyboard
up / down keys would always reset the cursor to the end of the number.
There also were intermittent "---" values appearing in the value when
drag-changing the plug.

This seems to come from
`PlugValueWidget.__updateFromValuesInBackgroundPreCall()` where the
value is set to `[]` before setting the real value.

Fixes #5335