BackgroundTask : Warn if ScriptNode can't be found
Not being able to find it can lead to very hard to diagnose bugs (see 67aea86), so we need to know any time that happens. This means updating a few tests that weren't using a ScriptNode at all. Technically, several of the tests could have been updated to pass `subject = None` instead, because they're not relying on graph edits causing cancellation. But I think it's better to update them to show the more common real-world use case instead.