SConstruct : Fix include order on Windows
MSVC seems to search include dirs in the order they are found on the command line, regardless of whether they are `/external:I` (system) or `/I` (regular) includes. This contrasts with GCC, which always searches regular includes before system includes. And unfortunately, the default SCons build commands place `$CXXFLAGS` before the include flags, so MSVC would find Gaffer's installed headers before it found the headers in the source tree, which are the ones we want. This reorders the build commands so that sanity can prevail.