SceneView : Fix cancellation of Grid's SceneGadget
Because the `GafferScene::Grid` node was just held as a member variable, it wasn't associated with any `ScriptNode` and therefore the computes done by the SceneView weren't cancellable by edits to the node or its inputs. By parenting everything into the `SceneView` child hierarchy we make a ScriptNode accessible to BackgroundTask and cancellation can be performed as intended. This is the way we should be implementing all bolt-in extensions to Views, so I've added a todo to that effect (fortunately none of the other extensions currently have cancellation problems).