Arnold config : Add default color manager
This translates Gaffer's context-based OCIO config into an Arnold `color_manager_ocio`, taking effect only if the scene doesn't already specify a color manager.

We wanted to do this as part of #5215, but couldn't because we needed to support Arnold 7.1, and that didn't have a modern enough OCIO to include the built-in ACES configs. Arnold 7.2 has a sufficiently modern OCIO.