OptionQueryUI : Fix check for existing queries
We were comparing the query plug names against the option name we wanted to query. But the plug names are just `query0`, `query1` etc, so the comparison is meaningless. In fact, the `plugName` argument to `__addQuery` had no affect on the outcome of the function, so I've simply removed it, and the code that was trying to sanitise names by replacing `:` with `_`.

I ran into this while looking for places that were avoiding `:` in plug names, since we allow that now anyway.